[#1] refactor: 차트 데이터 조작 로직 분리
